Organize the project directory and use English comments.
