+++ src/main/java/dev/latvian/kubejs/mekanism/custom/module/KubeJSModuleCallback.java
@@ -1,6 +1,6 @@
 package dev.latvian.kubejs.mekanism.custom.module;
 
-import dev.latvian.kubejs.mekanism.custom.CustomInterface;
+import dev.latvian.kubejs.mekanism.util.KJSInterfaceAddon;
 import mekanism.api.functions.TriConsumer;
 import mekanism.api.gear.ICustomModule;
 import mekanism.api.gear.IHUDElement;
@@ -37,15 +37,15 @@ public class KubeJSModuleCallback {
 
     public BiFunction<IModule<KubeJSModuleData>, ItemStack, Component> getModeScrollComponentCallback;
 
-    public CustomInterface.KQuintConsumer<IModule<KubeJSModuleData>, Player, ItemStack, Integer, Boolean, Void> changeModeCallback;
+    public KJSInterfaceAddon.KQuintConsumer<IModule<KubeJSModuleData>, Player, ItemStack, Integer, Boolean, Void> changeModeCallback;
 
-    public CustomInterface.KTriConsumer<IModule<KubeJSModuleData>, ItemStack, Consumer<NestedRadialMode>, Void> addRadialModesCallback;
+    public KJSInterfaceAddon.KTriConsumer<IModule<KubeJSModuleData>, ItemStack, Consumer<NestedRadialMode>, Void> addRadialModesCallback;
 
-    public CustomInterface.KTriConsumer<IModule<KubeJSModuleData>, ItemStack, RadialData<? extends IRadialMode>, ? extends IRadialMode> getModeCallback;
+    public KJSInterfaceAddon.KTriConsumer<IModule<KubeJSModuleData>, ItemStack, RadialData<? extends IRadialMode>, ? extends IRadialMode> getModeCallback;
 
-    public CustomInterface.KQuintConsumer<IModule<KubeJSModuleData>, Player, ItemStack, RadialData<? extends IRadialMode>, ? extends IRadialMode, Boolean> setModeCallback;
+    public KJSInterfaceAddon.KQuintConsumer<IModule<KubeJSModuleData>, Player, ItemStack, RadialData<? extends IRadialMode>, ? extends IRadialMode, Boolean> setModeCallback;
 
-    public CustomInterface.KQuadConsumer<IModule<KubeJSModuleData>, Player, LivingEntity, InteractionHand, InteractionResult> onInteractCallback;
+    public KJSInterfaceAddon.KQuadConsumer<IModule<KubeJSModuleData>, Player, LivingEntity, InteractionHand, InteractionResult> onInteractCallback;
 
     public BiFunction<IModule<KubeJSModuleData>, BlockSource, ICustomModule.ModuleDispenseResult> onDispenseCallback;
